barrelcard refactors
