+++ frontend/src/app/globals.css
@@ -48,17 +48,17 @@
   }
 
   @keyframes dots {
-    0%, 20% {
+    0% {
       content: '.';
     }
-    40% {
+    33.3% {
       content: '..';
     }
-    60% {
+    66.6% {
       content: '...';
     }
-    90%, 100% {
-      content: '';
+    100% {
+      content: '.';
     }
   }
 }
