+++ frontend/src/components/barrels/library/barrels-grid.tsx
@@ -5,7 +5,7 @@ import { Button, Card, Pagination } from "@nextui-org/react";
 import { useState } from "react";
 import { useSession } from "next-auth/react";
 import AddBarrelCard from "./add-barrel-card";
-import ModalBarrelCreateForm from "./modal-barrel-create-form";
+import BarrelCreateFormModal from "./barrel-create-form-modal";
 interface BarrelsListProps {
   publicBarrels: BarrelWithData[],
   privateBarrels: BarrelWithData[]
@@ -21,12 +21,8 @@ export default function BarrelsGrid({ publicBarrels, privateBarrels }: BarrelsLi
   const ITEMS_PER_PAGE_PRIVATE = 4;
   const ITEMS_PER_PAGE_PUBLIC = 8;
 
-  console.log("privateBarrels.length: " + privateBarrels.length)
-
   const numPrivateBarrelPages = Math.ceil(privateBarrels.length / ITEMS_PER_PAGE_PRIVATE);
-  console.log("private barrel pages: " + numPrivateBarrelPages)
   const numPublicBarrelPages = Math.ceil(publicBarrels.length / ITEMS_PER_PAGE_PUBLIC);
-  console.log("public barrel pages: " + numPublicBarrelPages)
 
   const handlePrivatePageChange = (newPage: number) => {
     setPrivateBarrelsPage(newPage);
@@ -36,15 +32,13 @@ export default function BarrelsGrid({ publicBarrels, privateBarrels }: BarrelsLi
     setPublicBarrelsPage(newPage);
   };
 
-  function getBarrelsForPublicPage() {
+  function barrelsForPublicPage() {
     const startItem = ((publicBarrelsPage - 1) * ITEMS_PER_PAGE_PUBLIC);
     const pageBarrels = publicBarrels.slice(startItem, startItem + ITEMS_PER_PAGE_PUBLIC);
     return pageBarrels;
   }
 
-  function getBarrelsForPrivatePage() {
-    console.log(privateBarrelsPage);
-
+  function barrelsForPrivatePage() {
     const startItem = ((privateBarrelsPage - 1) * ITEMS_PER_PAGE_PRIVATE);
     const pageBarrels = privateBarrels.slice(startItem, startItem + ITEMS_PER_PAGE_PRIVATE);
     return pageBarrels;
@@ -54,11 +48,11 @@ export default function BarrelsGrid({ publicBarrels, privateBarrels }: BarrelsLi
     {status === "authenticated" && <>
       <div className="mx-5 mb-5 text-xl">Private Barrels</div>
       <div className="grid gap-0 grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5">
-        {privateBarrelsPage === 1 && <ModalBarrelCreateForm>
+        {privateBarrelsPage === 1 && <BarrelCreateFormModal>
           <AddBarrelCard />
-        </ModalBarrelCreateForm>}
+        </BarrelCreateFormModal>}
 
-        {getBarrelsForPrivatePage().map((barrel) => (
+        {barrelsForPrivatePage().map((barrel) => (
           <BarrelCard type={BarrelType.private} key={"card" + barrel.id} barrel={barrel} color={"#BBBBEE"} />
         ))}
       </div>
@@ -72,7 +66,7 @@ export default function BarrelsGrid({ publicBarrels, privateBarrels }: BarrelsLi
       (<>
         <div className="mx-5 mb-5 text-xl">Public Barrels</div>
         <div className="grid gap-0 grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
-          {getBarrelsForPublicPage().map((barrel) => (
+          {barrelsForPublicPage().map((barrel) => (
             <BarrelCard type={BarrelType.public} key={"card" + barrel.id} barrel={barrel} color={"#BBBBEE"} />
           ))}
         </div>
